cash_machine/pyproject.toml
[tool.poetry]
name = "cash-machine"
version = "0.1.0"
description = "A test task for a Python developer"
authors = ["Arslan Semenov <arslan.semenoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
jinja2 = "^3.1.4"
asyncpg = "^0.29.0"
pdfkit = "^1.0.0"
qrcode = "^7.4.2"
pydantic-settings = "^2.4.0"
orjson = "^3.10.7"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
